refactor(FilePreview): Organize better code
