+++ common/src/utils/img_dimensions_preview.rs
@@ -1,0 +1,3 @@
+pub const IMAGE_MAX_WIDTH: &str = "80vw";
+
+pub const IMAGE_MAX_HEIGHT: &str = "80vh";
