+++ ui/src/utils/mod.rs
@@ -16,7 +16,6 @@ pub mod format_timestamp;
 pub mod get_drag_event;
 pub mod get_font_sizes;
 pub mod keyboard;
-pub mod lifecycle;
 pub mod verify_valid_paths;
 
 pub type EvalProvider = Rc<dyn Fn(&str) -> Result<UseEval, EvalError>>;
